Adapt C tests to the new printf behavior (mpfr 4.2.1) (#31)
Since printf now prints sign of NaNs, cleanup temporary values in some
C tests to avoid NaN being printed with the sign of the temporary
value stored in the output variable.

On the ML side, we doesn't handle output values, they are created on
demand and are unique. Sign bit is therefore clean.